split tracker out of models
