+++ .secrets.baseline
@@ -162,7 +162,7 @@
         "filename": "apps/bbsync/tests/test_integration.py",
         "hashed_secret": "3c3b274d119ff5a5ec6c1e215c1cb794d9973ac1",
         "is_verified": false,
-        "line_number": 93,
+        "line_number": 94,
         "is_secret": false
       }
     ],
@@ -182,7 +182,7 @@
         "filename": "apps/trackers/tests/test_integration.py",
         "hashed_secret": "3c3b274d119ff5a5ec6c1e215c1cb794d9973ac1",
         "is_verified": false,
-        "line_number": 237,
+        "line_number": 238,
         "is_secret": false
       }
     ],
@@ -192,7 +192,7 @@
         "filename": "apps/trackers/tests/test_save.py",
         "hashed_secret": "3c3b274d119ff5a5ec6c1e215c1cb794d9973ac1",
         "is_verified": false,
-        "line_number": 70,
+        "line_number": 71,
         "is_secret": false
       }
     ],
@@ -394,7 +394,7 @@
         "filename": "osidb/tests/endpoints/flaws/test_update_trackers.py",
         "hashed_secret": "3c3b274d119ff5a5ec6c1e215c1cb794d9973ac1",
         "is_verified": false,
-        "line_number": 81,
+        "line_number": 82,
         "is_secret": false
       }
     ],
@@ -404,7 +404,7 @@
         "filename": "osidb/tests/endpoints/test_affects.py",
         "hashed_secret": "3c3b274d119ff5a5ec6c1e215c1cb794d9973ac1",
         "is_verified": false,
-        "line_number": 73,
+        "line_number": 74,
         "is_secret": false
       }
     ],
@@ -434,7 +434,7 @@
         "filename": "osidb/tests/test_mixins.py",
         "hashed_secret": "3c3b274d119ff5a5ec6c1e215c1cb794d9973ac1",
         "is_verified": false,
-        "line_number": 776,
+        "line_number": 777,
         "is_secret": false
       }
     ],
@@ -467,5 +467,5 @@
       }
     ]
   },
-  "generated_at": "2024-09-16T18:19:17Z"
+  "generated_at": "2024-09-18T10:59:28Z"
 }
