+++ apps/bbsync/tests/test_integration.py
@@ -35,11 +35,12 @@ pytestmark = pytest.mark.integration
 @pytest.fixture(autouse=True)
 def enable_bbsync_env_var(monkeypatch) -> None:
     import apps.bbsync.mixins as mixins
+    import osidb.dmodels.tracker as tracker
     import osidb.models as models
 
     monkeypatch.setattr(mixins, "SYNC_TO_BZ", True)
     monkeypatch.setattr(models, "SYNC_FLAWS_TO_BZ", True)
-    monkeypatch.setattr(models, "SYNC_TRACKERS_TO_BZ", True)
+    monkeypatch.setattr(tracker, "SYNC_TRACKERS_TO_BZ", True)
 
 
 class TestBBSyncIntegration:
