+++ apps/bbsync/tests/test_query.py
@@ -7,7 +7,8 @@ from django.utils import timezone
 from apps.bbsync.query import FlawBugzillaQueryBuilder
 from apps.workflows.workflow import WorkflowModel
 from osidb.core import generate_acls
-from osidb.models import Affect, Flaw, FlawComment, FlawSource, Snippet, Tracker
+from osidb.dmodels.tracker import Tracker
+from osidb.models import Affect, Flaw, FlawComment, FlawSource, Snippet
 from osidb.tests.factories import (
     AffectFactory,
     FlawCommentFactory,
