+++ apps/exploits/serializers.py
@@ -2,7 +2,8 @@ from rest_framework import serializers
 
 from apps.exploits.models import EPSS, Exploit
 from osidb.dmodels import PsModule
-from osidb.models import Affect, Flaw, Tracker
+from osidb.dmodels.tracker import Tracker
+from osidb.models import Affect, Flaw
 
 
 class ExploitsSimpleSerializer(serializers.ModelSerializer):
