+++ apps/exploits/tests/test_api.py
@@ -3,7 +3,8 @@ import json
 import pytest
 
 from apps.exploits.helpers import store_or_update_exploits
-from osidb.models import Affect, Tracker
+from osidb.dmodels.tracker import Tracker
+from osidb.models import Affect
 from osidb.tests.factories import (
     AffectFactory,
     FlawFactory,
