+++ apps/trackers/common.py
@@ -13,8 +13,9 @@ from apps.trackers.constants import (
 from apps.trackers.exceptions import TrackerCreationError
 from collectors.bzimport.constants import BZ_URL
 from osidb.dmodels import PsModule, PsUpdateStream
+from osidb.dmodels.tracker import Tracker
 from osidb.helpers import cve_id_comparator
-from osidb.models import Flaw, Tracker
+from osidb.models import Flaw
 
 
 class TrackerQueryBuilder:
