+++ apps/trackers/save.py
@@ -1,7 +1,8 @@
 """
 common tracker save funtionality
 """
-from osidb.models import Flaw, Tracker
+from osidb.dmodels.tracker import Tracker
+from osidb.models import Flaw
 
 from .bugzilla.save import TrackerBugzillaSaver
 from .exceptions import BTSException, UnsupportedTrackerError
