+++ apps/trackers/tests/test_bugzilla.py
@@ -11,7 +11,8 @@ from apps.bbsync.exceptions import ProductDataError
 from apps.bbsync.tests.factories import BugzillaComponentFactory, BugzillaProductFactory
 from apps.sla.tests.test_framework import load_sla_policies
 from apps.trackers.bugzilla.query import TrackerBugzillaQueryBuilder
-from osidb.models import Affect, Impact, Tracker
+from osidb.dmodels.tracker import Tracker
+from osidb.models import Affect, Impact
 from osidb.tests.factories import (
     AffectFactory,
     FlawFactory,
