+++ apps/trackers/tests/test_common.py
@@ -5,7 +5,8 @@ import pytest
 from django.utils import timezone
 
 from apps.trackers.common import TrackerQueryBuilder
-from osidb.models import Affect, Flaw, Tracker
+from osidb.dmodels.tracker import Tracker
+from osidb.models import Affect, Flaw
 from osidb.tests.factories import (
     AffectFactory,
     FlawFactory,
