+++ apps/trackers/tests/test_jira.py
@@ -28,6 +28,7 @@ from apps.trackers.jira.query import (
 )
 from apps.trackers.models import JiraProjectFields
 from apps.trackers.tests.factories import JiraProjectFieldsFactory
+from osidb.dmodels.tracker import Tracker
 from osidb.models import (
     CVSS,
     Affect,
@@ -36,7 +37,6 @@ from osidb.models import (
     FlawCVSS,
     Impact,
     PsUpdateStream,
-    Tracker,
 )
 from osidb.tests.factories import (
     AffectCVSSFactory,
