+++ collectors/errata/core.py
@@ -12,7 +12,7 @@ from requests_gssapi import HTTPSPNEGOAuth
 from collectors.framework.models import CollectorMetadata
 from osidb.core import set_user_acls
 from osidb.dmodels.erratum import Erratum
-from osidb.models import Tracker
+from osidb.dmodels.tracker import Tracker
 
 from ..utils import BACKOFF_KWARGS, fatal_code
 from .constants import (
