+++ collectors/errata/tests/test_core.py
@@ -3,7 +3,8 @@ from django.utils import timezone
 
 from collectors.framework.models import CollectorMetadata
 from osidb.dmodels.erratum import Erratum
-from osidb.models import Affect, Tracker
+from osidb.dmodels.tracker import Tracker
+from osidb.models import Affect
 from osidb.tests.factories import (
     AffectFactory,
     ErratumFactory,
