+++ collectors/framework/tests/test_core.py
@@ -3,7 +3,8 @@ from celery.schedules import crontab
 from django.utils import timezone
 
 from collectors.framework.models import CollectorMetadata, collector
-from osidb.models import Affect, Flaw, Tracker
+from osidb.dmodels.tracker import Tracker
+from osidb.models import Affect, Flaw
 
 pytestmark = pytest.mark.unit
