+++ collectors/jiraffe/tasks.py
@@ -6,7 +6,8 @@ from celery.schedules import crontab
 from celery.utils.log import get_task_logger
 
 from collectors.framework.models import collector
-from osidb.models import Flaw, Tracker
+from osidb.dmodels.tracker import Tracker
+from osidb.models import Flaw
 
 from .collectors import JiraTaskCollector, JiraTrackerCollector, MetadataCollector
 from .constants import (
