+++ collectors/jiraffe/tests/test_convertors.py
@@ -5,7 +5,8 @@ import pytest
 
 from collectors.jiraffe.convertors import JiraTrackerConvertor
 from collectors.jiraffe.core import JiraQuerier
-from osidb.models import Affect, Tracker
+from osidb.dmodels.tracker import Tracker
+from osidb.models import Affect
 from osidb.sync_manager import JiraTrackerLinkManager
 from osidb.tests.factories import (
     AffectFactory,
