+++ osidb/dmodels/erratum.py
@@ -3,8 +3,8 @@ import uuid
 from django.core.exceptions import ObjectDoesNotExist
 from django.db import models
 
+from osidb.dmodels.tracker import Tracker
 from osidb.mixins import TrackingMixin, TrackingMixinManager
-from osidb.models import Tracker
 
 
 class ErratumManager(TrackingMixinManager):
