+++ osidb/signals.py
@@ -8,8 +8,9 @@ from django.utils import timezone
 from jira import JIRA
 
 from osidb.dmodels import Profile
+from osidb.dmodels.tracker import Tracker
 from osidb.helpers import get_env
-from osidb.models import Affect, AffectCVSS, Flaw, FlawCVSS, Tracker
+from osidb.models import Affect, AffectCVSS, Flaw, FlawCVSS
 
 logger = logging.getLogger(__name__)
