+++ osidb/tests/endpoints/flaws/test_update_trackers.py
@@ -5,7 +5,8 @@ from django.utils import timezone
 from django.utils.timezone import datetime
 from rest_framework import status
 
-from osidb.models import Affect, Tracker
+from osidb.dmodels.tracker import Tracker
+from osidb.models import Affect
 from osidb.tests.factories import (
     AffectFactory,
     FlawFactory,
