+++ osidb/tests/endpoints/test_affects.py
@@ -3,7 +3,8 @@ from unittest.mock import patch
 import pytest
 from rest_framework import status
 
-from osidb.models import Affect, AffectCVSS, Tracker
+from osidb.dmodels.tracker import Tracker
+from osidb.models import Affect, AffectCVSS
 from osidb.tests.factories import (
     AffectCVSSFactory,
     AffectFactory,
